requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
numpy>=1.24.0
httpx>=0.25.0
torch>=2.0.0
transformers>=4.30.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0 